requirements.txt
pyMeterBus
paho-mqtt
RPi.GPIO